minor bug

+++ transform_wrappers.py
@@ -29,10 +29,10 @@ class DensityTransform:
         graphs_adj_matrices = list(np.squeeze(graph_adj_tensors.numpy()))
         if alpha > 0:
             edited_adj_matrices = [nx.adjacency_matrix(densify(nx.convert_matrix.from_numpy_matrix(g), alpha)).todense() for g in graphs_adj_matrices]
-            return torch.unsqueeze(torch.from_numpy(np.asarray(edited_adj_matrices)), -1)
+            return torch.unsqueeze(torch.from_numpy(np.asarray(edited_adj_matrices).astype(float)), -1)
         else: 
             edited_adj_matrices = [nx.adjacency_matrix(sparsify(nx.convert_matrix.from_numpy_matrix(g), 0 - alpha)).todense() for g in graphs_adj_matrices]
-            return torch.unsqueeze(torch.from_numpy(np.asarray(edited_adj_matrices)), -1)
+            return torch.unsqueeze(torch.from_numpy(np.asarray(edited_adj_matrices).astype(float)), -1)
             
 
 ## transform via adding edges
